% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_timestamp_create.R
\name{action_timestamp_create}
\alias{action_timestamp_create}
\title{Create a timestamp}
\usage{
action_timestamp_create()
}
\description{
Format the timestamp so that it can be used in log file
names, too.
}
\keyword{internal}
