% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_id_statement.R, R/add_item_statement.R,
%   R/add_statement.R
\name{add_id_statement}
\alias{add_id_statement}
\alias{add_item_statement}
\alias{add_statement}
\title{Add statements to an item}
\usage{
add_id_statement(
  qid,
  pid,
  o,
  wikibase_type = "external-id",
  wikibase_api_url = "https://reprexbase.eu/demowiki/api.php",
  csrf
)

add_item_statement(
  qid,
  pid,
  o,
  wikibase_type = "item",
  wikibase_api_url = NULL,
  csrf
)

add_statement(
  qid,
  pid,
  o,
  wikibase_type = c("item", "string", "numeric", "external-id"),
  wikibase_api_url = "https://reprexbase.eu/demowiki/api.php",
  csrf
)
}
\arguments{
\item{qid}{The QID of the item in the Wikibase instance.}

\item{pid}{The PID of the property to be added.}

\item{o}{The object of the statement (either a QID, string, or number).}

\item{wikibase_type}{One of \code{"item"}, \code{"string"} (or \code{"external-id"}), or \code{"numeric"}.}

\item{wikibase_api_url}{The full URL of the Wikibase API (must end with \code{api.php}).}

\item{csrf}{The CSRF token of your session, received with
\code{\link{get_csrf}}.}
}
\value{
A data.frame describing the created statement.
}
\description{
See
\link[https://www.wikidata.org/w/api.php?action=help&modules=wbcreateclaim]{MediaWiki
API help}

Adds a claim (statement) to a Wikibase item via \code{wbcreateclaim}.
See
\link[https://www.wikidata.org/w/api.php?action=help&modules=wbcreateclaim]{MediaWiki
API help}
}
