% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item.R
\name{create_item}
\alias{create_item}
\title{Create an item}
\usage{
create_item(
  label,
  description,
  language,
  equivalence_property = NA_character_,
  equivalence_id = NA_character_,
  classification_property = NA_character_,
  classification_id = NA_character_,
  wikibase_api_url = NULL,
  data_curator = NULL,
  log_file_name = NULL,
  csrf = NULL,
  wikibase_session = NULL
)
}
\arguments{
\item{label}{A language-specific, human-readable name for the Wikibase entity
(item, property, or Lexeme). Used when setting or retrieving the entity’s
label in the specified language, for example, \code{"human"}.}

\item{description}{A short, language-specific textual summary of the Wikibase
entity (such as an item, property, or Lexeme). Descriptions help
disambiguate entities with similar labels and are not required to be unique
but are typically distinct within the same language. For example, \code{"A member of the
homo sapience species."}.}

\item{language}{A single language code that indicates the language of the
label and description, using BCP 47-compliant language tags (e.g., "en" for
English, "fr" for French).}

\item{equivalence_property}{An optional PID of a property already defined in
the same Wikibase instance that records the equivalence of this new
item with a property defined elsewhere, for example, on Wikidata or
CIDOC-CRM. Defaults to \code{NA_character_}; if left missing, no
equivalence relations is will be claimed.}

\item{equivalence_id}{The identifier that uniquely identifies this item among
another system's definitions. Defaults to
\code{NA_character_}; if left missing, no equivalence relations is will be
claimed.}

\item{classification_property}{The instance of, or subclass of, or superclass
of property. Defaults to \code{NA_character} when not used.}

\item{classification_id}{The QID of the class. Defaults to
\code{NA_character} when not used.}

\item{wikibase_api_url}{The full URL of the Wikibase API, which is the
address that the \code{wbdataset} R client sends requests to when
interacting with the knowledge base. For example,
\code{'https://reprexbase.eu/jekyll/api.php'}. The URL must end with
api.php. It is either given as a parameter or resolved from
\code{wikibase_session}.}

\item{data_curator}{The name of the data curator who runs the function and
creates the log file, created with \link[utils]{person}.
It is either given as a parameter or resolved from
\code{wikibase_session}. If no curator is given, then filled with
\code{person("Unknown", "Curator")}.}

\item{log_file_name}{An explicitly stated full path to a possible CSV log
file, defaults to \code{NULL}. If the value is \code{NULL}, no log file
will be created.}

\item{csrf}{The CSRF token of your session, received with
\code{\link{get_csrf}}. It is either given as a parameter or resolved
from \code{wikibase_session}.}

\item{wikibase_session}{An optional named list of default values to reuse
across multiple function calls. If any of the main parameters (such as
\code{language}, \code{data_curator}, \code{log_file_name},
\code{equivalence_propeert}, \code{classification_property}
\code{wikibase_api_url}, or \code{csrf}) are missing from the function
call, their values will be taken from this list if available. This is
useful in interactive workflows or scripts where the same context is
reused.}
}
\value{
Returns a \code{\link[dataset]{dataset_df}} object.
The columns are:\cr
\describe{
\item{\code{action}}{\code{create_item}}
\item{\code{id_on_target}}{The new Property Identifier (PID) on the targeted Wikibase.}
\item{\code{label}}{The property label}
\item{\code{description}}{The description label}
\item{\code{language}}{The language code of the label}
\item{\code{datatype}}{Defaults to \code{item}}
\item{\code{wikibase_api_url}}{The MediaWiki API URL where the new property is created}
\item{\code{equivalence_property}}{The PID that connects an equivalence ID to the property}
\item{\code{equivalence_id}}{The ID of an equivalent item defined elsewhere}
\item{\code{classification_property}}{The PID that connects the item to a superclass, or class.}
\item{\code{classification_id}}{The QID of a class, subclass or superclass.}
\item{\code{success}}{TRUE if successfully created, FALSE if there was an error}
\item{\code{comment}}{A summary of the error messages(s), if success is FALSE}
\item{\code{time}}{The time when the action started}
\item{\code{logfile}}{The name of the CSV logfile}
}
}
\description{
Creates an item entity on a Wikibase instance with a single
language labelling and description. New labels and descriptions can be added
in further language later. This is a wrapper for
\href{https://www.wikidata.org/w/api.php?action=help&modules=wbeditentity}{MediaWiki
action=wbeditentity}
}
\details{
Optionally, the function adds an equivalent item to this newly
created item.
}
\examples{
\dontrun{
# Define a reusable session
my_session <- list(
  language = "en",
  csrf = get_csrf("username", "password",
  wikibase_api_url = "https://example.org/w/api.php"),
  data_curator = person("Jane", "Doe"),
  wikibase_api_url = "https://example.org/w/api.php"
)
create_item(
   label = "biological object"
   language = "en"
   description = "This class comprises individual items of a material nature,
    which live, have lived or are natural products of or from living organisms.",
   equivalence_property = "P37"  # property for equivalent CIDOC definitions
   equivalence_id = "E20"  # CIDOC-CRM identifier of this concept
   wikibase_session = my_session)
}
}
