% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_property.R
\name{create_property}
\alias{create_property}
\title{Create a property}
\usage{
create_property(
  label,
  description,
  language,
  datatype,
  equivalence_property,
  equivalaence_id,
  wikibase_api_url,
  csrf
)
}
\arguments{
\item{description}{A single description, for example,  \code{"Length in time."}.}

\item{language}{A single language code, for example, \code{"en"}.}

\item{datatype}{A single character string defining the datatype of this
property, for example, \code{"quantity".}}

\item{equivalence_property}{An optional PID of a property already defined in
the same Wikibase instance that records the equivalence of this new property
with a property defined elsewhere, for example, on Wikidata or CIDOC-CRM.}

\item{wikibase_api_url}{For example, \code{'https://reprexbase.eu/demowiki/api.php'}.}

\item{csrf}{The CSRF token of your session, received with \code{\link{get_csrf}}.}

\item{equivalence_id}{An optional PID or other identifier in Wikibase
that records the equivalence of the property. An external ID.}
}
\value{
Currently returns a data.frame, this should be a dataset. It has four
columns, \code{default_label},
\code{pid_on_source} containing the PID of the property on the source instance,
\code{pid_on_target} containing the new PID on the target Wikibase instance,
\code{success}, a logical value if the copying was
successful.
}
\description{
Creates a property entity on a Wikibase instance with a single language
labelling and description. New labels and descriptions can be added
in further languages later.
}
\details{
Optionally, the function adds an equivalent property to this newly created
property.
}
\examples{
\dontrun{
# Only works with authentication
your_csrf <- get_csrf(
   username  = your_username,
   password = your_password,
   wikibase_api_url = "https://reprexbase.eu/demowiki/api.php"
   )

create_property(
   label = "duration"
   language = "en"
   description = "Length of the process in time"
   datatype = "quantity",
   csrf = your_csrf)
}

}
