% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_claim.R
\name{get_claims}
\alias{get_claims}
\title{Get claims of an item}
\usage{
get_claims(
  qid = "Q528626",
  property = "P625",
  wikibase_api_url = "https://www.wikidata.org/w/api.php",
  csrf = NULL
)
}
\arguments{
\item{qid}{The QID of the item.}

\item{property}{The property for which the claim is required.}

\item{wikibase_api_url}{The URL of the Wikibase API.}

\item{csrf}{A response csrf received with \code{\link{get_csrf}}.}
}
\description{
Get the claims (statements) related to an item.
}
