% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csrf.R
\name{get_csrf}
\alias{get_csrf}
\alias{get_csrf_token}
\title{Get CSRF}
\usage{
get_csrf(username, password, wikibase_api_url)

get_csrf_token(csrf)
}
\arguments{
\item{username}{Your bot username.}

\item{password}{Your bot password}

\item{wikibase_api_url}{The API URL for the Wikibase instance.}

\item{csrf}{The csrf object returned by \code{\link{get_csrf}}.}
}
\value{
A list, the CSRF response

The CSRF response token.
}
\description{
Get a Cross-Site Request Forgery (CSRF) response
}
