% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item.R
\name{get_item}
\alias{get_item}
\title{Get item}
\usage{
get_item(
  qid,
  language,
  url = "https://www.wikidata.org/w/api.php",
  creator = person("Jane", "Doe"),
  title = "Dataset title"
)
}
\arguments{
\item{qid}{A single QID or a vector of QIDs on a Wikibase instance (or Wikidata itself.)}

\item{language}{descr}

\item{url}{Defaults to \code{"https://www.wikidata.org/w/api.php"}.}

\item{creator}{Author}

\item{title}{Title}
}
\description{
Get item
}
\examples{
qid = "Q42"

}
