% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_propety_definition.R
\name{get_property_definition}
\alias{get_property_definition}
\title{Get property definition}
\usage{
get_property_definition(
  pid,
  language = c("en", "nl", "hu"),
  wikibase_api_url = "https://www.wikidata.org/w/api.php",
  return_type = "JSON",
  csrf = NULL
)
}
\arguments{
\item{pid}{The PID of the property in the Wikibase instance (or Wikidata
itself).}

\item{language}{Defaults to \code{c("en", "nl", "hu")}. A character string of
the languages in which the users wants to receive the labels and
descriptions of the property. The vector of languages must use \href{https://en.wikipedia.org/wiki/IETF_language_tag}{BCP
47}-compliant language tags (e.g., "en" for English, and "hu"
for Hungarian.)}

\item{wikibase_api_url}{The full URL of the Wikibase API, which is the
address that the \code{wbdataset} R client sends requests to when
interacting with the knowledge base. In this case it defaults to
\code{'https://www.wikidata.org/w/api.php'}, Wikidata itself, where no CSRF
is needed.}

\item{return_type}{Defaults to \code{"data.frame"} that is suitable for
receiving the information in stand-alone use. The \code{"JSON"} passes on a
JSON string in the format that you may need it in further Wikibase API
calls.}

\item{csrf}{The CSRF token of your session, received with
\code{\link{get_csrf}}, not needed if
\code{wikibase_api_url="https://www.wikidata.org/w/api.php"}. Defaults to
\code{NULL}.}
}
\value{
A data.frame of the \code{PID} with the labels and descriptions of
the property in the selected languages. Alternatively, when
\code{return_type="JSON"}, the same data prepared for use in a subsequent
API call, for example, to copy these contents into a new Wikibase instance.
}
\description{
Receive the label and description of a property on the basis of
its PID from a Wikibase instance. It will not add further statements about
the property.
}
\details{
Currently the language has a choose a default, \code{"en"}, for
cases where the user-chosen language return empty labels and descriptions.
This feature may be elaborated or changed later. The function receives
aliases (alternative labels), too, but does not return them; the format of
aliases needs to be decided in view of how other functions will use them,
as aliases themselves can break the tidiness of the returned data.
}
\examples{
# Receive a data.frame for further use
get_property_definition(pid = "P2047", return_type = "data.frame")

# Receive JSON for copying with wbeditidentiy
get_property_definition(pid = "P2047", language = c("en", "hu"))
}
