% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-here.R
\name{here}
\alias{here}
\title{Here}
\arguments{
\item{...}{Path components below the project root, can be empty.
Each argument should be a string containing one or more path components
separated by a forward slash "/"}
}
\value{
The project root is established.
}
\description{
See \code{here::\link[here:here]{here}} for details.
}
\keyword{internal}
