% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_df_not_empty}
\alias{is_df_not_empty}
\title{Is an object a non-empty data.frame?}
\usage{
is_df_not_empty(df)
}
\arguments{
\item{df}{An object}
}
\value{
A boolean, \code{TRUE} if the object is a data.frame with at least one
row.
}
\description{
Is an object a non-empty data.frame?
}
\keyword{internal}
