% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_qid_is_pid.R
\name{is_pid}
\alias{is_pid}
\title{Check if a string looks like a PID}
\usage{
is_pid(x)
}
\arguments{
\item{x}{A string}
}
\value{
A logical vector, \code{TRUE} where the string has a "P\link{:digits:}" structure,
FALSE otherwise.
}
\description{
Check if a string looks like a PID
}
