% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_response_success}
\alias{is_response_success}
\title{API response success}
\usage{
is_response_success(response)
}
\arguments{
\item{response}{A response from MediaWiki Action API, type response.}
}
\description{
API response success
}
\keyword{internal}
