% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_csrf}
\alias{is_valid_csrf}
\title{Check if a CSRF token appears valid}
\usage{
is_valid_csrf(csrf)
}
\arguments{
\item{csrf}{A single character string representing the token.}
}
\value{
Logical \code{TRUE} or \code{FALSE}.
}
\description{
Validates the basic structure of a MediaWiki-style CSRF token.
}
\keyword{internal}
