% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_wikibase_datatype.R
\name{is_valid_wikibase_datatype}
\alias{is_valid_wikibase_datatype}
\title{Check if a datatype is a valid Wikibase datatype}
\usage{
is_valid_wikibase_datatype(x)
}
\arguments{
\item{x}{A character vector of datatypes to validate. Non-character inputs
will be coerced using \code{as.character()}.}
}
\value{
A logical vector indicating whether each input value is a valid
Wikibase datatype.
}
\description{
This function checks whether a given character vector of
datatypes are valid according to the Wikibase Media API specification.
}
\examples{
# Valid inputs
is_valid_wikibase_datatype("wikibase-item")
is_valid_wikibase_datatype(c("string", "url", "external-id"))

# Invalid input (not lowercase, not API-compatible)
\dontrun{
# FALSE — should be "external-id"
is_valid_wikibase_datatype("External ID")
}
}
\seealso{
\href{https://www.mediawiki.org/wiki/Wikibase/DataModel/JSON#Datatypes}{Wikibase JSON DataModel documentation}
}
