% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_identity_datastring_create.R
\name{item_identity_datastring_create}
\alias{item_identity_datastring_create}
\title{Create a JSON for wbeditentity items}
\usage{
item_identity_datastring_create(labels_list, descriptions_list)
}
\arguments{
\item{labels_list}{A list property labels organised by language.}

\item{descriptions_list}{A list of descriptions organised by language.}
}
\value{
A JSON string for the wbeditentity API call
}
\description{
Creates a JSON string of the label(s), descriptions(s) in various languages,
adds the datatype. Required to define a property on Wikibase.
}
\keyword{internal}
