% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_claim.R
\name{left_join_column}
\alias{left_join_column}
\title{Join a new column by property}
\usage{
left_join_column(
  ds,
  property,
  wikibase_api_url = "https://www.wikidata.org/w/api.php",
  silent = FALSE
)
}
\arguments{
\item{ds}{A dataset object that contains the observations (QIDs)}

\item{property}{The property}

\item{wikibase_api_url}{The URL of the Wikibase API.}

\item{csrf}{A response csrf}
}
\description{
Get the claims (statements) related to an item.
}
