% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property_identity_datastring_create.R
\name{property_identity_datastring_create}
\alias{property_identity_datastring_create}
\title{Create a JSON for wbeditentity properties}
\usage{
property_identity_datastring_create(labels_list, descriptions_list, datatype)
}
\arguments{
\item{labels_list}{A list property labels organised by language.}

\item{descriptions_list}{A list of descriptions organised by language.}

\item{datatype}{An allowed datatype of the wbeditidentity, for example,
\code{"quantity"}.}
}
\value{
A JSON string for the wbeditentity API call
}
\description{
Creates a JSON string of the label(s), descriptions(s) in various languages,
adds the datatype. Required to define a property on Wikibase.
}
\keyword{internal}
