% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_wikibase_entities.R
\name{search_wikibase_entities}
\alias{search_wikibase_entities}
\title{Search Wikibase Entities}
\usage{
search_wikibase_entities(
  search_term,
  language = "en",
  wikibase_api_url = "https://www.wikidata.org/w/api.php",
  type,
  csrf = NULL
)
}
\arguments{
\item{search_term}{A character string representing the term to search for.}

\item{language}{A character string specifying the language code (e.g., "en").}

\item{wikibase_api_url}{A character string providing the URL of the Wikibase
API endpoint.}

\item{type}{A character string indicating the type of entity to search for:
"item" or "property". Defaults to "item".}
}
\value{
A list containing search results from the Wikibase API.
}
\description{
This internal function interfaces with the Wikibase API to
search for entities matching a given search term and language.
}
\keyword{internal}
