% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_wikibase_session.R
\name{validate_wikibase_session}
\alias{validate_wikibase_session}
\title{Validate a wikibase_session object}
\usage{
validate_wikibase_session(wikibase_session)
}
\arguments{
\item{session}{A list, typically created with \code{\link{new_wikibase_session}}.}
}
\value{
\code{TRUE} if the session is valid, otherwise an informative error.
}
\description{
Checks that the session object is a named list and its components are valid for reuse.
}
\keyword{internal}
