% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wikidata_countries_df.R, R/data.R
\docType{data}
\name{wikidata_countries_df}
\alias{wikidata_countries_df}
\title{Countries on Wikidata}
\format{
A data frame with 4 rows and 4 columns:
\describe{
\item{qid}{QID of the item.}
\item{label}{Name label of the item.}
\item{description}{description of the item.}
\item{language}{Language of the title and description.}
...
}

An object of class \code{wbdataset} (inherits from \code{dataset_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\source{
\url{https://www.wikidata.org/wiki/Wikidata:Main_Page}
}
\usage{
wikidata_countries_df

wikidata_countries_df
}
\description{
A small dataset of countries, including sovereign states and autonomous
parts of federal states taken from Wikidata.
}
\keyword{datasets}
