% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_id_statement.R, R/add_statement.R
\name{add_id_statement}
\alias{add_id_statement}
\alias{add_statement}
\alias{add_wikibase_url_statement}
\title{Add statements}
\usage{
add_id_statement(
  qid,
  pid,
  o,
  wikibase_type = "external-id",
  wikibase_api_url = "https://reprexbase.eu/jekyll/api.php",
  csrf
)

add_statement(
  qid,
  pid,
  o,
  wikibase_type,
  wikibase_api_url = "https://reprexbase.eu/demowiki/api.php",
  csrf
)

add_wikibase_url_statement(
  qid,
  pid,
  wikibase_url,
  wikibase_api_url = "https://reprexbase.eu/demowiki/api.php",
  csrf_token
)
}
\arguments{
\item{qid}{The QID of the item in the Wikibase instance that you use.}

\item{pid}{The PID of the equivalent Wikidata (or reference Wikibase) URI.}

\item{o}{The object of the semantic statement.}

\item{wikibase_type}{A \code{'item'}, \code{'numeric'}, or \code{'string'}.}

\item{wikibase_api_url}{\code{\link{"https://reprexbase.eu/demowiki/api.php"}}.}

\item{csrf}{The CSRF token of your session, received with \code{\link{get_csrf}}.}

\item{wikidata_uri}{The QID on Wikidata (or a reference Wikibase).}
}
\value{
A data.frame with four columns:
\code{id} containing the statement ID, \code{qid} with the QID,
\code{o} with the PID, and \code{p} with the  wikidata_uri.
}
\description{
See \link[https://www.wikidata.org/w/api.php?action=help&modules=wbcreateclaim]{MediaWiki API help}
}
