% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_existing_item.R
\name{check_existing_item}
\alias{check_existing_item}
\title{Check for Existing Items in Wikibase}
\usage{
check_existing_item(
  search_term,
  language = "en",
  equivalence_property = NA_character_,
  equivalence_id = NA_character_,
  classification_property = NA_character_,
  classification_id = NA_character_,
  action = "create_item",
  log_file_name = NA_character_,
  data_curator = person("Unknown", "Person"),
  wikibase_api_url = "https://www.wikidata.org/w/api.php",
  csrf = NULL
)
}
\arguments{
\item{search_term}{A character string representing the label to search for
in the Wikibase. For example, \code{"Estonian National Museum"}.}

\item{language}{A character string specifying the language code of the label,
adhering to BCP 47 standards (e.g., \code{"en"} for English). Defaults to \code{"en"}.}

\item{equivalence_property}{An optional PID of a property already defined in
the same Wikibase instance that records the equivalence of this new
item with a property defined elsewhere, for example, on Wikidata or
CIDOC-CRM. Defaults to \code{NA_character_}; if left missing, no
equivalence relations is will be claimed.}

\item{equivalence_id}{The identifier that uniquely identifies this item among
another system's definitions. Defaults to
\code{NA_character_}; if left missing, no equivalence relations is will be
claimed.}

\item{classification_property}{The instance of, or subclass of, or superclass
of property. Defaults to \code{NA_character} when not used.}

\item{classification_id}{The QID of the class. Defaults to
\code{NA_character} when not used.}

\item{action}{A character string indicating the action being performed.
Defaults to \code{"create_item"}.}

\item{log_file_name}{A character string specifying the name of the log file.
Defaults to \code{NA_character_}.}

\item{data_curator}{An object of class \code{person} representing the data curator.
Defaults to \code{person("Unknown", "Person")}.}

\item{wikibase_api_url}{The full URL of the Wikibase API, which is the
address that the \code{wbdataset} R client sends requests to when
interacting with the knowledge base. For example,
\code{'https://reprexbase.eu/demowiki/api.php'}. The URL must end with
api.php.}

\item{csrf}{The CSRF token of your session, received with
\code{\link{get_csrf}}.}
}
\value{
A \code{dataset_df} object containing information about the matching item(s),
including action performed, item ID, label, description, language, and other
metadata. Returns \code{NULL} if no matching items are found.

A data.frame or NULL.
}
\description{
This function searches for existing items in a specified Wikibase instance
that match a given label in a specified language. It helps prevent duplicate
item creation by identifying existing matches.
}
\details{
The function interfaces with the Wikibase API's \code{wbsearchentities} action to
perform a search based on the provided label (\code{search_term}) and language.
It returns information about matching items, facilitating the management of
data consistency within the Wikibase.
See:
\href{https://www.wikidata.org/w/api.php?action=help&modules=wbsearchentities}{MediaWiki
action=wbsearchentities}.
}
\examples{
check_existing_item(
  search_term = "Estonian National Museum",
  language = "en",
  wikibase_api_url = "https://www.wikidata.org/w/api.php",
  csrf = NULL
)
}
