% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_existing_property.R
\name{check_existing_property}
\alias{check_existing_property}
\title{Check if a label already has an property.}
\usage{
check_existing_property(
  action = "create_property",
  search_term,
  language,
  action_timestamp = NULL,
  equivalence_property = NA_character_,
  equivalence_id = NA_character_,
  classification_property = NA_character_,
  classification_id = NA_character_,
  log_file_name = NA_character_,
  data_curator = person("Unknown", "Person"),
  wikibase_api_url,
  csrf
)
}
\arguments{
\item{action}{Defaults to \code{"create_property"}.}

\item{search_term}{A label in the given language, for example, "Estonia".}

\item{language}{A single language code, for example, \code{"en"}.}

\item{equivalence_property}{An optional PID of a property already defined in
the same Wikibase instance that records the equivalence of this new
property with a property defined elsewhere, for example, on Wikidata or
CIDOC-CRM. Defaults to \code{NA_character_}; if left missing, no
equivalence relations is will be claimed.}

\item{equivalence_id}{An optional PID or other identifier in Wikibase that
records the equivalence of the property. An external ID. Defaults to
\code{NA_character_}; if left missing, no equivalence relations is will be
claimed.}

\item{data_curator}{The name of the data curator who runs the function and
creates the log file, created with \link[utils]{person}.}

\item{wikibase_api_url}{For example,
\code{'https://reprexbase.eu/demowiki/api.php'}.}

\item{csrf}{The CSRF token of your session, received with
\code{\link{get_csrf}}.}
}
\value{
A data.frame or NULL.
}
\description{
Avoid failed writing attempts by checking if a label already matches an item.
}
\details{
A wrapper around
\href{https://www.wikidata.org/w/api.php?action=help&modules=wbsearchentities}{MediaWiki
action=wbsearchentities}.
}
