% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_property.R
\name{copy_property}
\alias{copy_property}
\title{Copy an property}
\usage{
copy_property(
  pid_on_wikidata,
  wikidata_pid_property = "P2",
  languages = c("en", "hu"),
  wikibase_api_url = "https://reprexbase.eu/jekyll/api.php",
  csrf
)
}
\arguments{
\item{pid_on_wikidata}{The PID of the property on Wikidata to be copied to your Wikibase.}

\item{wikidata_pid_property}{The PID in Wikibase that records the equivalent Wikidata
PID as an external ID.}

\item{wikibase_api_url}{For example, \code{'https://reprexbase.eu/demowiki/api.php'}.}

\item{csrf}{The CSRF token of your session, received with \code{\link{get_csrf}}.}

\item{language}{A vector of languages codes, for example, \code{c("en", "et")}.}
}
\description{
Copy an property
}
