% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_wikidata_item.R,
%   R/copy_wikidata_property.R
\name{copy_wikidata_item}
\alias{copy_wikidata_item}
\alias{copy_wikidata_items}
\alias{copy_wikidata_properties}
\title{Copy a Wikidata item(s)}
\usage{
copy_wikidata_item(
  qid_on_wikidata = "Q4",
  qid_equivalence_property = "P35",
  language = c("en", "nl", "hu"),
  wikibase_api_url = "https://reprexbase.eu/jekyll/api.php",
  data_curator = NULL,
  log_path = tempdir(),
  csrf,
  wikibase_session = NULL
)

copy_wikidata_items(
  qid_on_wikidata,
  qid_equivalence_property,
  language,
  wikibase_api_url,
  data_curator,
  log_path,
  csrf
)

copy_wikidata_properties(
  pid_on_wikidata,
  pid_equivalence_property,
  language,
  wikibase_api_url,
  data_curator,
  log_path,
  csrf,
  wikibase_session = NULL
)
}
\arguments{
\item{qid_on_wikidata}{The QID of the item to be copied to your Wikibase. A
single valid QID or a vector of several QIDs.}

\item{qid_equivalence_property}{The QID in Wikibase that records the
equivalent Wikidata QID as an external ID.}

\item{language}{A vector of language codes, for example, \code{c("en",
"et")}.}

\item{wikibase_api_url}{For example,
\code{'https://reprexbase.eu/demowiki/api.php'}.}

\item{data_curator}{The name of the data curator who runs the function and
creates the log file, created with \link[utils]{person}.}

\item{log_path}{A path to save the log file. Defaults to the return value of
\code{\link{tempdir()}}.}

\item{csrf}{The CSRF token of your session, received with
\code{\link{get_csrf}}.}

\item{wikibase_session}{An optional list that contains any of the values of
parameters \code{qid_equivalence_property}, \code{language},
\code{wikibase_api_url}, \code{data_curator},\code{log_path} and
\code{csrf} (for repeated use in a session.)}
}
\value{
Returns a dataset_df object. The columns are:
\itemize{
\item{"rowid"}{ A row identifier. }
\item{"action"}{ The create_item() function name. }
\item{"id_on_target"}{ The new item identifier (QID) on the targeted Wikibase.}
\item{"label"}{ The propery label}
\item{"description"}{ The description label}
\item{"language"}{ The language code of the label.}
\item{"datatype"}{ The datatype of the property, for example, \code{string}}
\item{"wikibase_api_url"}{ The MediaWiki API URL where the new property is created.}
\item{"equivalence_property"}{ The PID that connects an equivalence ID to the property.}
\item{"equivalence_id"}{ The ID of an equivalent property defined elsewhere.}
\item{"success"}{ TRUE if successfully created, FALSE if there was an error.}
\item{"comment"}{ A summary of the error messages(s), if success is FALSE.}
\item{"time"}{ The time when the action started.}
\item{"logfile"}{ The name of the CSV logfile.}
}
The number of rows corresponds to the length of the qid_on_wikidata vector.
}
\description{
This code will copy (an) item label(s) and description(s) from
Wikidata to a new instance. It should work between instances, but the
authentication to copy from a password protected instance is not yet coded.
The function is more specific than create_item, because this one creates
items that exist in a similarly structured Wikibase instance, such as
Wikidata.
}
\details{
This function is slightly different from create_item. That function
creates a new item that may not have an equivalent item on another Wikibase
instance, but it may well have an equivalent item in another database. \cr
In this function, we use \code{qid_equivalence_property} for equivalence.
In the more general create function we use \code{equivalence_item}, because
we may use different identifiers. Similarly, the \code{qid_on_wikidata}
replaces the more general \code{equivalence_id}, because we must use QID
for identification in an other Wikibase instance.
}
\keyword{internal}
