% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_wikidata_property.R
\name{copy_wikidata_property}
\alias{copy_wikidata_property}
\title{Copy a Wikidata property or properties}
\usage{
copy_wikidata_property(
  pid_on_wikidata,
  pid_equivalence_property = "P2",
  language = c("en", "hu"),
  wikibase_api_url = "https://reprexbase.eu/jekyll/api.php",
  data_curator = NULL,
  log_path = tempdir(),
  csrf,
  wikibase_session = NULL
)
}
\arguments{
\item{pid_on_wikidata}{The PID of the property on Wikidata to be copied to
your Wikibase.It can be one or more valid PIDs. (Only works with
non-authenticated sources, this should be changed.)}

\item{pid_equivalence_property}{The PID in Wikibase that records the
equivalent Wikidata PID as an external ID.}

\item{language}{A vector of language codes, for example, \code{c("en",
"et")}.}

\item{wikibase_api_url}{For example,
\code{'https://reprexbase.eu/demowiki/api.php'}.}

\item{data_curator}{The name of the data curator who runs the function and
creates the log file, created with \link[utils]{person}.}

\item{log_path}{A path to save the log file. Defaults to the return value of
\code{\link{tempdir()}}.}

\item{csrf}{The CSRF token of your session, received with
\code{\link{get_csrf}}.}

\item{wikibase_session}{An optional list that contains any of the values of
parameters \code{qid_equivalence_property}, \code{language},
\code{wikibase_api_url}, \code{data_curator},\code{log_path} and
\code{csrf} (for repeated use in a session.)}

\item{classification_property}{The instance of, or subclass of, or superclass
of property. Defaults to \code{NA_character} when not used.}

\item{classification_id}{The QID of the class. Defaults to
\code{NA_character} when not used.}
}
\value{
Returns a dataset_df object. The columns are:
\itemize{
\item{"rowid"}{ A row identifier. }
\item{"action"}{ The create_property() function name. }
\item{"id_on_target"}{ The new Property Identifier (PID) on the targeted Wikibase.}
\item{"label"}{ The propery label}
\item{"description"}{ The description label}
\item{"language"}{ The language code of the label.}
\item{"datatype"}{ The datatype of the property, for example, \code{string}}
\item{"wikibase_api_url"}{ The MediaWiki API URL where the new property is created.}
\item{"equivalence_property"}{ The PID that connects an equivalence ID to the property.}
\item{"equivalence_id"}{ The ID of an equivalent property defined elsewhere.}
\item{"classification_property"}{ Not applicable for properties.}
\item{"classification_id"}{ Not applicable for properties.}
\item{"success"}{ TRUE if successfully created, FALSE if there was an error.}
\item{"comment"}{ A summary of the error messages(s), if success is FALSE.}
\item{"time"}{ The time when the action started.}
\item{"logfile"}{ The name of the CSV logfile.}
}
The number of rows corresponds to the length of the qid_on_wikidata vector.
}
\description{
This code will copy one or multiple property label(s) and
description(s) from Wikidata to a new instance. It should work between
instances, but the authentication to copy from a password protected
instance is not yet coded. The function is more specific than
\code{\link{create_property}}, because this one creates properties that
exist in a similarly structured Wikibase instance, such as Wikidata.
}
\details{
This function is slightly different from
\code{\link{create_property}}. That function creates a new property that
may not have an equivalent property on another Wikibase instance, but it
may well have an equivalent property in another graph system. \cr In this
function, we use \code{pid_equivalence_property} for equivalence. In the
more general create function we use \code{equivalence_property}, because we
may use different identifiers. Similarly, the \code{pid_on_wikidata}
replaces the more general \code{equivalence_id}, because we must use PID
for identification in an other Wikibase instance.
}
