% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item.R
\name{create_item}
\alias{create_item}
\title{Create a property}
\usage{
create_item(
  label,
  description,
  language,
  equivalence_property = NA_character_,
  equivalence_id = NA_character_,
  classification_property = NA_character_,
  classification_id = NA_character_,
  wikibase_api_url,
  data_curator = NULL,
  log_path = tempdir(),
  csrf,
  wikibase_session
)
}
\arguments{
\item{description}{A single description, for example, \code{"A member of the
homo sapience species."}.}

\item{language}{A single language code, for example, \code{"en"}.}

\item{equivalence_property}{An optional PID of a property already defined in
the same Wikibase instance that records the equivalence of this new
item with a property defined elsewhere, for example, on Wikidata or
CIDOC-CRM. Defaults to \code{NA_character_}; if left missing, no
equivalence relations is will be claimed.}

\item{equivalence_id}{The identifier that uniquely identifies this item among
another system's defintions. Defaults to
\code{NA_character_}; if left missing, no equivalence relations is will be
claimed.}

\item{classification_property}{The instance of, or subclass of, or superclass
of property. Defaults to \code{NA_character} when not used.}

\item{classification_id}{The QID of the class. Defaults to
\code{NA_character} when not used.}

\item{wikibase_api_url}{For example,
\code{'https://reprexbase.eu/demowiki/api.php'}.}

\item{data_curator}{The name of the data curator who runs the function and
creates the log file, created with \link[utils]{person}.}

\item{log_path}{A path to save the log file. Defaults to the return value of
\code{\link{tempdir()}}.}

\item{csrf}{The CSRF token of your session, received with
\code{\link{get_csrf}}.}

\item{wikibase_session}{An optional list that contains any of the values of
\code{language},
\code{wikibase_api_url}, \code{data_curator},\code{log_path} and
\code{csrf} (for repeated use in a session.)}
}
\value{
Currently returns a data.frame, this should be a dataset. The columns
are:
\itemize{
\item{"action"}{ create_property}
\item{"id_on_target"}{ The new Property Identifier (PID) on the targeted Wikibase.}
\item{"label"}{ The propery label}
\item{"description"}{ The description label}
\item{"language"}{ The language code of the label}
\item{"datatype"}{ Defaults to \code{item}}
\item{"wikibase_api_url"}{ The MediaWiki API URL where the new property is created}
\item{"equivalence_property"}{ The PID that connects an equivalence ID to the property}
\item{"equivalence_id"}{ The ID of an equivalent item defined elsewhere}
\item{"classification_property"}{ The PID that connects the item to a superclass, or class.}
\item{"classification_id"}{ The QID of a class, subclass or superclass.}
\item{"success"}{ TRUE if successfully created, FALSE if there was an error}
\item{"comment"}{ A summary of the error messages(s), if success is FALSE}
\item{"time"}{ The time when the action started}
\item{"logfile"}{ The name of the CSV logfile}
}
}
\description{
Creates an item entity on a Wikibase instance with a single
language labelling and description. New labels and descriptions can be added
in further language later. This is a wrapper for
\href{https://www.wikidata.org/w/api.php?action=help&modules=wbeditentity}{MediaWiki
action=wbeditentity}
}
\details{
Optionally, the function adds an equivalent item to this newly
created item.
}
\examples{
\dontrun{
# Only works with authentication
your_csrf <- get_csrf(
   username  = your_username,
   password = your_password,
   wikibase_api_url = "https://reprexbase.eu/demowiki/api.php"
   )

create_item(
   label = "biological object"
   language = "en"
   description = "This class comprises individual items of a material nature,
    which live, have lived or are natural products of or from living organisms.",
   equivalence_property = "P37"  # property for equivalent CIDOC definitions
   equivalence_id = "E20"  # CIDOC-CRM identifier of this concept
   csrf = your_csrf)
}
}
