% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_property.R
\name{create_property}
\alias{create_property}
\title{Create a property}
\usage{
create_property(
  label,
  description,
  language,
  datatype,
  equivalence_property = NA_character_,
  equivalence_id = NA_character_,
  wikibase_api_url,
  data_curator = NULL,
  log_file_name = NULL,
  csrf,
  wikibase_session = NULL
)
}
\arguments{
\item{label}{A language-specific, human-readable name for the Wikibase entity
(item, property, or Lexeme). Used when setting or retrieving the entity’s
label in the specified language, for example, \code{"duration"}.}

\item{description}{A short, language-specific textual summary of the Wikibase
entity (such as an item, property, or Lexeme). Descriptions help
disambiguate entities with similar labels and are not required to be unique
but are typically distinct within the same language. For example, \code{"Length in
  time."}.}

\item{language}{A single language code that indicates the language of the
label and description, using BCP 47-compliant language tags (e.g., "en" for
English, "fr" for French).}

\item{datatype}{A single character string defining the datatype of this
property, for example, \code{"string"}, \code{"wikibase-item"},
\code{"quantity"}, \code{"commonsMedia"}, \code{"url"}, \code{"external-id"},
\code{"globe-coordinate"}.}

\item{equivalence_property}{An optional PID of a property already defined in
the same Wikibase instance that records the equivalence of this new
property with a property defined elsewhere, for example, on Wikidata or
CIDOC-CRM. Defaults to \code{NA_character_}; if left missing, no
equivalence relations is will be claimed.}

\item{equivalence_id}{An optional PID or other identifier in Wikibase that
records the equivalence of the property. An external ID. Defaults to
\code{NA_character_}; if left missing, no equivalence relations is will be
claimed.}

\item{wikibase_api_url}{The full URL of the Wikibase API, which is the
address that the \code{wbdataset} R client sends requests to when
interacting with the knowledge base. For example,
\code{'https://reprexbase.eu/demowiki/api.php'}. The URL must end with
api.php.}

\item{data_curator}{The name of the data curator who runs the function and
creates the log file, created with \link[utils]{person}.}

\item{log_file_name}{An explicitly stated full path to a possible CSV log
file, defaults to \code{NULL}. If the value is \code{NULL}, no log file
will be created.}

\item{csrf}{The CSRF token of your session, received with
\code{\link{get_csrf}}.}

\item{wikibase_session}{An optional list that contains any of the values of
\code{language},
\code{wikibase_api_url}, \code{data_curator}, and
\code{csrf} (for repeated use in a session.)}
}
\value{
Returns a \code{\link[dataset]{dataset_df}} object.
The columns are:\cr
\describe{
\item{\code{action}}{\code{create_property}}
\item{\code{id_on_target}}{The new Property Identifier (PID) on the targeted Wikibase.}
\item{\code{label}}{The propery label}
\item{\code{description}}{The description label}
\item{\code{language}}{The language code of the label.}
\item{\code{datatype}}{The datatype of the property, for example, \code{string}}
\item{\code{wikibase_api_url}}{The MediaWiki API URL where the new property is created.}
\item{\code{equivalence_property}}{The PID that connects an equivalence ID to the property.}
\item{\code{equivalence_id}}{The ID of an equivalent property defined elsewhere.}
\item{\code{classification_property}}{Not applicable for properties.}
\item{\code{classification_id}}{Not applicable for properties.}
\item{\code{success}}{TRUE if successfully created, FALSE if there was an error.}
\item{\code{comment}}{A summary of the error messages(s), if success is FALSE.}
\item{\code{time}}{The time when the action started.}
\item{\code{logfile}}{The name of the CSV logfile.}
}
}
\description{
Creates a property entity on a Wikibase instance with a single
language labelling and description. New labels and descriptions can be added
in further language later. This is a wrapper for
\href{https://www.wikidata.org/w/api.php?action=help&modules=wbeditentity}{MediaWiki
action=wbeditentity}
}
\details{
Optionally, the function adds an equivalent property to this newly
created property.
}
\examples{
\dontrun{
# Only works with authentication
your_csrf <- get_csrf(
   username  = your_username,
   password = your_password,
   wikibase_api_url = "https://reprexbase.eu/demowiki/api.php"
   )

create_property(
   label = "duration"
   language = "en"
   description = "Length of the process in time"
   datatype = "quantity",
   csrf = your_csrf)
}

}
