% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_property.R
\name{create_property}
\alias{create_property}
\title{Create a property}
\usage{
create_property(
  label,
  description,
  language,
  datatype,
  equivalence_property = NA_character_,
  equivalence_id = NA_character_,
  wikibase_api_url,
  data_curator = NULL,
  log_path = tempdir(),
  csrf
)
}
\arguments{
\item{description}{A single description, for example,  \code{"Length in
time."}.}

\item{language}{A single language code, for example, \code{"en"}.}

\item{datatype}{A single character string defining the datatype of this
property, for example, \code{"quantity".}}

\item{equivalence_property}{An optional PID of a property already defined in
the same Wikibase instance that records the equivalence of this new
property with a property defined elsewhere, for example, on Wikidata or
CIDOC-CRM. Defaults to \code{NA_character_}; if left missing, no
equivalence relations is will be claimed.}

\item{equivalence_id}{An optional PID or other identifier in Wikibase that
records the equivalence of the property. An external ID. Defaults to
\code{NA_character_}; if left missing, no equivalence relations is will be
claimed.}

\item{wikibase_api_url}{For example,
\code{'https://reprexbase.eu/demowiki/api.php'}.}

\item{data_curator}{The name of the data curator who runs the function and
creates the log file, created with \link[utils]{person}.}

\item{log_path}{A path to save the log file. Defaults to the return value of
\code{\link{tempdir()}}.}

\item{csrf}{The CSRF token of your session, received with
\code{\link{get_csrf}}.}
}
\value{
Currently returns a data.frame, this should be a dataset. The columns
are:
\itemize{
\item{"action"}{ create_property}
\item{"id_on_target"}{ The new Property Identifier (PID) on the targeted Wikibase.}
\item{"label"}{ The propery label}
\item{"description"}{ The description label}
\item{"language"}{ The language code of the label.}
\item{"datatype"}{ The datatype of the property, for example, \code{string}}
\item{"wikibase_api_url"}{ The MediaWiki API URL where the new property is created.}
\item{"equivalence_property"}{ The PID that connects an equivalence ID to the property.}
\item{"equivalence_id"}{ The ID of an equivalent property defined elsewhere.}
\item{"success"}{ TRUE if successfully created, FALSE if there was an error.}
\item{"comment"}{ A summary of the error messages(s), if success is FALSE.}
\item{"time"}{ The time when the action started.}
\item{"logfile"}{ The name of the CSV logfile.}
}
}
\description{
Creates a property entity on a Wikibase instance with a single
language labelling and description. New labels and descriptions can be added
in further languages later. This is a wrapper for
\href{https://www.wikidata.org/w/api.php?action=help&modules=wbeditentity}{MediaWiki
action=wbeditentity}
}
\details{
Optionally, the function adds an equivalent property to this newly
created property.
}
\examples{
\dontrun{
# Only works with authentication
your_csrf <- get_csrf(
   username  = your_username,
   password = your_password,
   wikibase_api_url = "https://reprexbase.eu/demowiki/api.php"
   )

create_property(
   label = "duration"
   language = "en"
   description = "Length of the process in time"
   datatype = "quantity",
   csrf = your_csrf)
}

}
