% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item.R
\name{get_item}
\alias{get_item}
\title{Get items}
\usage{
get_item(
  qid,
  language,
  prefix = "http://www.wikidata.org/entity/",
  wikibase_api_url = "https://www.wikidata.org/w/api.php",
  creator = person("Jane", "Doe"),
  title = "Dataset title"
)
}
\arguments{
\item{qid}{A single QID or a vector of QIDs on a Wikibase instance (or Wikidata itself.)}

\item{language}{A character vector of language codes, for example,
\code{c("en", "nl", "hu")}.}

\item{prefix}{The prefix to use before the QID, for example, defaults to
\code{"http://www.wikidata.org/entity/"}.}

\item{wikibase_api_url}{Defaults to \code{"https://www.wikidata.org/w/api.php"}.}

\item{creator}{The creator (author) of the dataset.}

\item{title}{The title of the dataset.}
}
\value{
A dataset with the QIDs, labels, description, and the language codes
of the labels and descriptions.
}
\description{
Get item definitions by QID from a Wikibase instance or Wikidata.
}
\examples{
get_item("Q42", language = c("en", "nl"))

}
