% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wikidata_item.R
\name{get_wikidata_item}
\alias{get_wikidata_item}
\title{Get item definitions}
\usage{
get_wikidata_item(
  qid_on_wikidata,
  language,
  prefix = "http://www.wikidata.org/entity/",
  wikibase_api_url = "https://www.wikidata.org/w/api.php",
  data_curator = NULL,
  title = "Dataset title"
)
}
\arguments{
\item{qid_on_wikidata}{A single qid_on_wikidata or a vector of qid_on_wikidatas on a Wikibase instance (or
Wikidata itself.)}

\item{language}{Defaults to \code{c("en", "nl", "hu")}. A character string of
the languages in which the users wants to receive the labels and
descriptions of the property. The vector of languages must use \href{https://en.wikipedia.org/wiki/IETF_language_tag}{BCP
47}-compliant language tags (e.g., "en" for English, "nl" for Dutch and "hu"
for Hungarian.)}

\item{prefix}{The prefix to use before the qid_on_wikidata, for example, defaults to
\code{"http://www.wikidata.org/entity/"}.}

\item{wikibase_api_url}{Defaults to
\code{"https://www.wikidata.org/w/api.php"}.}

\item{data_curator}{The name of the data curator who runs the function and
creates the log file, created with \link[utils]{person}.}

\item{title}{The title of the dataset.}
}
\value{
A Returns a \code{\link[dataset]{dataset_df}} object with the
qid_on_wikidatas, labels, description, and the language codes of the labels
and descriptions.
}
\description{
Get item definitions by qid_on_wikidata from a Wikibase instance or
Wikidata.
}
\examples{
get_wikidata_item("Q42", language = c("en", "nl"))
}
