% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wikidata_item.R
\name{get_wikidata_item}
\alias{get_wikidata_item}
\title{Retrieve Wikidata item metadata (labels and descriptions)}
\usage{
get_wikidata_item(
  qid_on_wikidata,
  language,
  prefix = "http://www.wikidata.org/entity/",
  wikibase_api_url = "https://www.wikidata.org/w/api.php",
  data_curator = NULL,
  title = "Dataset title",
  fallback_language = "en"
)
}
\arguments{
\item{qid_on_wikidata}{A character vector of Wikidata QIDs (e.g.,
\code{"Q42"}).}

\item{wikibase_api_url}{The full URL of the Wikibase API endpoint (must end
with \code{api.php}).}

\item{fallback_language}{A BCP 47 language code (e.g., "en") used as a fallback
when no label is available in the requested language(s). If a label is missing
in one or more specified languages, this fallback will be used instead when available.
Defaults to \code{"en"}.}
}
\value{
A data frame with one row per QID, including columns \code{qid},
\code{label}, and \code{description}, as well as optional language-specific
metadata.
}
\description{
Retrieves basic metadata (labels and descriptions) for one or
more Wikidata items, given their QIDs. Supports multiple languages and
returns results in a tidy format.
}
\details{
This function queries the Wikidata API to retrieve the \code{label}
and \code{description} for each item specified in \code{qid_on_wikidata}.
The metadata is returned in a consistent tabular format and can be used for
display, annotation, or joining with other datasets.

By default, the function retrieves metadata in multiple languages,
including English, French, German, Italian, and Spanish.
}
\examples{
\dontrun{
get_wikidata_item("Q42", language = c("hu"), fallback_language = "en")
}
}
\seealso{
\code{\link{get_claim}} to retrieve claims (properties) for Wikidata
items.
}
