% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wikidata_item.R
\name{get_wikidata_item}
\alias{get_wikidata_item}
\title{Retrieve Wikidata item metadata (labels and descriptions)}
\usage{
get_wikidata_item(
  qid_on_wikidata,
  language,
  prefix = "http://www.wikidata.org/entity/",
  wikibase_api_url = "https://www.wikidata.org/w/api.php",
  data_curator = NULL,
  title = "Dataset title",
  fallback_language = "en"
)
}
\arguments{
\item{qid_on_wikidata}{A character vector of Wikidata QIDs (e.g.,
\code{"Q42"}).}

\item{language}{Defaults to \code{c("en")}. A character string of
the languages in which the users wants to receive the labels and
descriptions of the property. The vector of languages must use \href{https://en.wikipedia.org/wiki/IETF_language_tag}{BCP
47}-compliant language tags (e.g., "en" for English, and "hu"
for Hungarian.)}

\item{prefix}{The prefix to be used as a namespace for downloaded items.}

\item{wikibase_api_url}{The full URL of the Wikibase API endpoint (must end
with \code{api.php}).}

\item{data_curator}{The name of the data curator who runs the function and
creates the log file, created with \link[utils]{person}.
It is either given as a parameter or resolved from
\code{wikibase_session}. If no curator is given, then filled with
\code{person("Unknown", "Curator")}.}

\item{title}{The title of the returned (log) dataset.}

\item{fallback_language}{A BCP 47 language code (e.g., "en") used as a fallback
when no label is available in the requested language(s). If a label is missing
in one or more specified languages, this fallback will be used instead when available.
Defaults to \code{"en"}.}
}
\value{
A data frame with one row per QID, including columns \code{qid},
\code{label}, and \code{description}, as well as optional language-specific
metadata.
}
\description{
Retrieves basic metadata (labels and descriptions) for one or
more Wikidata items, given their QIDs. Supports multiple languages and
returns results in a tidy format.
}
\details{
This function queries the Wikidata API to retrieve the \code{label}
and \code{description} for each item specified in \code{qid_on_wikidata}.
The metadata is returned in a consistent tabular format and can be used for
display, annotation, or joining with other datasets.

By default, the function retrieves metadata in multiple languages,
including English, French, German, Italian, and Spanish.
}
\examples{
\dontrun{
get_wikidata_item("Q42", language = c("hu"), fallback_language = "en")
}
}
\seealso{
\code{\link{get_claim}} to retrieve claims (properties) for Wikidata
items.
}
