% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_search_term_ambiguity.R
\name{handle_search_term_ambiguity}
\alias{handle_search_term_ambiguity}
\title{Handle Search Term Ambiguity}
\usage{
handle_search_term_ambiguity(
  search_results,
  search_term,
  language,
  strategy = c("return_null", "return_first"),
  csrf = NULL
)
}
\arguments{
\item{search_results}{A list of search results from the Wikibase API.}

\item{search_term}{The original search term used in the query.}

\item{language}{The language code used in the search.}

\item{strategy}{If \code{return_null} than a NULL value is returned if
there are ambiguous results, if \code{return_first} then the first result
is returned.}
}
\value{
A single search result or NULL if ambiguity cannot be resolved.
}
\description{
Resolves ambiguity in search results by selecting a preferred
match or returning NULL.
}
\keyword{internal}
