% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_column.R
\name{left_join_column}
\alias{left_join_column}
\title{Join a new column by property}
\usage{
left_join_column(
  ds,
  property,
  label = NULL,
  unit = NULL,
  definition = NULL,
  namespace = NULL,
  wikibase_api_url = "https://www.wikidata.org/w/api.php",
  silent = FALSE,
  csrf = NULL
)
}
\arguments{
\item{ds}{A dataset object that contains the observations (by QIDs).}

\item{property}{The property}

\item{label}{A short, human-readable description of the vector or \code{NULL}.}

\item{unit}{A character string of length one containing the unit of measure
or \code{NULL}.}

\item{definition}{A character string of length one containing a linked
definition or \code{NULL}.}

\item{namespace}{A namespace for individual observations or categories or
\code{NULL}.}

\item{wikibase_api_url}{The URL of the Wikibase API.}

\item{silent}{If the function should send messages about individual requests
to the API or not. Defaults to \code{FALSE}.}

\item{csrf}{A}
}
\description{
Get the claims (statements) related to an item.
}
